Moved content of app/dashboard/profile/types into store/profile.store.ts for cleaner imports.
