+++ packages/app/store/profile.store.ts
@@ -1,8 +1,38 @@
+import { Profile } from '@/types/model';
 import { generateErrorMessage, uploadImage } from '@/utils';
 import axios from 'axios';
 import { create } from 'zustand';
 import { CreateNewGroupFormType, ProfileFormType } from '../app/dashboard/profile/schema';
-import { Profile, ProfileActions, ProfileState, ProfileStore } from '../app/dashboard/profile/types';
+
+export type ProfileState = {
+  profile: Profile;
+  isProfileInitialized: boolean;
+  initializeProfileError?: string;
+  isUpdatingProfile: boolean;
+  profileUpdateError?: string;
+  isExitingGroup: boolean;
+  exitGroupError?: string;
+  isJoiningGroup: boolean;
+  joinGroupError?: string;
+  isCreatingGroup: boolean;
+  createGroupError?: string;
+  isUploadingAvatar?: boolean;
+  isSignedIn?: boolean;
+};
+
+export type ProfileActions = {
+  initializeProfile: () => Promise<void>;
+  uploadAvatar: (file: File, name?: string) => Promise<string | void>;
+  updateProfile: (profileData: ProfileFormType) => Promise<void>;
+  joinGroup: (groupId: string) => Promise<void>;
+  exitGroup: () => Promise<void>;
+  createGroup: (groupProfileData: CreateNewGroupFormType) => Promise<void>;
+  removeGroupAssociation: () => void;
+  setIsSignedIn: (isSignedIn: boolean) => void;
+  clearProfile: () => void;
+};
+
+export type ProfileStore = ProfileState & { actions: ProfileActions };
 
 const now = new Date();
 export const DEFAULT_PROFILE: Profile = {
